Refactor handleResponseText in ODataHandler
