+++ packages/handlers/odata/src/scalars.ts
@@ -1,0 +1,20 @@
+export const SCALARS = new Map<string, string>([
+  ['Edm.Binary', 'String'],
+  ['Edm.Stream', 'String'],
+  ['Edm.String', 'String'],
+  ['Edm.Int16', 'Int'],
+  ['Edm.Byte', 'Byte'],
+  ['Edm.Int32', 'Int'],
+  ['Edm.Int64', 'BigInt'],
+  ['Edm.Double', 'Float'],
+  ['Edm.Boolean', 'Boolean'],
+  ['Edm.Guid', 'GUID'],
+  ['Edm.DateTimeOffset', 'DateTime'],
+  ['Edm.Date', 'Date'],
+  ['Edm.TimeOfDay', 'String'],
+  ['Edm.Single', 'Float'],
+  ['Edm.Duration', 'ISO8601Duration'],
+  ['Edm.Decimal', 'Float'],
+  ['Edm.SByte', 'Byte'],
+  ['Edm.GeographyPoint', 'String'],
+]);
