+++ packages/handlers/odata/src/schema-util.ts
@@ -1,0 +1,13 @@
+import { EventEmitter } from 'events';
+
+export interface EntityTypeExtensions {
+  entityInfo: {
+    actualFields: string[];
+    navigationFields: string[];
+    identifierFieldName?: string;
+    identifierFieldTypeRef?: string;
+    isOpenType: boolean;
+  };
+  typeObj: any;
+  eventEmitter: EventEmitter;
+}
